Mort post-merge fixes
